+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Removed
+
+- Remove mainFields fix for webpack helpers as it is no longer necessary
+
 ## [0.2.0-alpha.3] - 2023-06-14
 
 ### Changed