+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Internal
+
+- Update to `@fastly/js-compute`@`3.6.2`
+
 ## [0.2.0] - 2023-09-19
 
 ### Fixed