+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Fixed
+
+- webpack-helpers: Correctly override loaders
+
 ### Internal
 
 - Update to `@fastly/js-compute`@`3.6.2`