+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.2.1] - 2023-10-12
+
 ### Added
 
 - Added README demo example in TypeScript
@@ -69,6 +71,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/fastly/compute-js-opentelemetry/compare/0.2.0...HEAD
-[0.2.0]: https://github.com/fastly/compute-js-opentelemetry/compare/0.1.0...v0.2.0
-[0.1.0]: https://github.com/fastly/compute-js-opentelemetry/releases/tag/0.1.0
+[unreleased]: https://github.com/fastly/compute-js-opentelemetry/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/fastly/compute-js-opentelemetry/compare/v0.2.0...v0.2.1
+[0.2.0]: https://github.com/fastly/compute-js-opentelemetry/compare/v0.1.1...v0.2.0
+[0.1.1]: https://github.com/fastly/compute-js-opentelemetry/compare/v0.1.0...v0.1.1
+[0.1.0]: https://github.com/fastly/compute-js-opentelemetry/releases/tag/v0.1.0